+++ src/cli_helper.rs
@@ -160,6 +160,9 @@ impl ActivityCli {
             .subcommand(SubCommand::with_name("continue").about("Continue a finished activity"))
             .subcommand(SubCommand::with_name("day").about("Display the current day as a timeline"))
             .subcommand(
+                SubCommand::with_name("week").about("Display the current week as a timeline"),
+            )
+            .subcommand(
                 SubCommand::with_name("delete")
                     .about("Delete activity")
                     .arg(Arg::with_name("id").required(true).help("activity id")),