+++ src/timeline.rs
@@ -1,13 +1,15 @@
 use ansi_term::{Color, Style};
 use anyhow::anyhow;
-use chrono::{DateTime, Duration, Local};
+use chrono::{DateTime, Datelike, Duration, Local, Timelike};
 use rtw::{Activity, ActivityId, DurationW};
 use std::cmp::max;
-use tbl::{Block, BlockRenderer, RenderBlock, Renderer};
+use tbl::{Block, BlockRenderer, Bound, RenderBlock, Renderer};
 
 type RGB = (u8, u8, u8);
 type Label = (String, RGB);
 type Interval = (ActivityId, Activity);
+
+const DEFAULT_TERMINAL_SIZE: usize = 90;
 
 struct ActivityRenderer {}
 
@@ -43,14 +45,19 @@ fn bounds(interval: &Interval) -> (f64, f64) {
     let (_, activity) = interval;
     let start_time: DateTime<Local> = activity.get_start_time().into();
     let stop_time: DateTime<Local> = activity.get_stop_time().into();
-    (start_time.timestamp() as f64, stop_time.timestamp() as f64)
+    (
+        start_time.num_seconds_from_midnight() as f64,
+        stop_time.num_seconds_from_midnight() as f64,
+    )
 }
 
+// label for activities
 fn label(interval: &Interval) -> Option<Label> {
     let (activity_id, activity) = interval;
     Some((activity.get_title(), color(*activity_id)))
 }
 
+// label for legend
 fn legend(interval: &Interval) -> Option<Label> {
     let (_activity_id, activity) = interval;
     let start_time: DateTime<Local> = activity.get_start_time().into();
@@ -65,28 +72,91 @@ fn legend(interval: &Interval) -> Option<Label> {
     ))
 }
 
-pub(crate) fn render(activities: &[Interval]) -> anyhow::Result<String> {
-    let (width, _height) = term_size::dimensions().ok_or_else(|| anyhow!("fqi"))?;
-    let total: Duration = activities
+// day total activities duration
+fn day_total(activities: &[Interval]) -> Duration {
+    activities
         .iter()
         .map(|(_, a)| {
             let duration: Duration = a.get_duration().into();
             duration
         })
-        .fold(Duration::seconds(0), |total, duration| total + duration);
-    let total = DurationW::from(total);
-    let total_string = total.to_string();
-    let right_padding = total_string.len() + 1; // +1 space
-    let available_length = max(0, width - right_padding as usize) as usize;
-    let legend = Renderer::new(activities, &bounds, &legend)
-        .with_renderer(&ActivityRenderer {})
-        .with_length(available_length)
-        .render()
-        .or_else(|_| Err(anyhow!("failed to create timeline")))?;
-    let timeline = Renderer::new(activities, &bounds, &label)
-        .with_renderer(&ActivityRenderer {})
-        .with_length(available_length)
-        .render()
-        .or_else(|_| Err(anyhow!("failed to create timeline")))?;
-    Ok(format!("{} total\n{} {}", legend, timeline, total_string))
+        .fold(Duration::seconds(0), |total, duration| total + duration)
+}
+
+// earliest and latest activity
+fn day_bounds(activities: &[Interval]) -> Bound {
+    let min_second: f64 = activities
+        .iter()
+        .map(|(_, a)| {
+            let start: DateTime<Local> = a.get_start_time().into();
+            start.num_seconds_from_midnight()
+        })
+        .min()
+        .unwrap_or(0) as f64;
+    let max_second = activities
+        .iter()
+        .map(|(_, a)| {
+            let stop: DateTime<Local> = a.get_stop_time().into();
+            stop.num_seconds_from_midnight()
+        })
+        .max()
+        .unwrap_or(86_400) as f64;
+    (min_second, max_second)
+}
+
+// min and max day
+fn days(activities: &[Interval]) -> (i32, i32) {
+    let min_day = activities
+        .iter()
+        .map(|(_, a)| {
+            let start: DateTime<Local> = a.get_start_time().into();
+            start.num_days_from_ce()
+        })
+        .min()
+        .unwrap_or(0);
+    let max_day = activities
+        .iter()
+        .map(|(_, a)| {
+            let stop: DateTime<Local> = a.get_stop_time().into();
+            stop.num_days_from_ce()
+        })
+        .max()
+        .unwrap_or(0);
+    (min_day, max_day)
+}
+
+pub(crate) fn render_days(activities: &[Interval]) -> anyhow::Result<Vec<String>> {
+    let (width, _height) = term_size::dimensions().unwrap_or((DEFAULT_TERMINAL_SIZE, 0));
+    let (min_second, max_second) = day_bounds(activities);
+    let (min_day, max_day) = days(activities);
+    let mut rendered: Vec<String> = vec![];
+    for day in min_day..=max_day {
+        let day_activities: Vec<Interval> = activities
+            .iter()
+            .filter(|&(_, a)| {
+                let start_time: DateTime<Local> = a.get_start_time().into();
+                start_time.num_days_from_ce() == day
+            })
+            .cloned()
+            .collect();
+        let total: DurationW = DurationW::from(day_total(day_activities.as_slice()));
+        let total_string = total.to_string();
+        let right_padding = total_string.len() + 1; // +1 space
+        let available_length = max(0, width - right_padding as usize) as usize;
+        let legend = Renderer::new(day_activities.as_slice(), &bounds, &legend)
+            .with_renderer(&ActivityRenderer {})
+            .with_length(available_length)
+            .with_boundaries((min_second, max_second))
+            .render()
+            .or_else(|_| Err(anyhow!("failed to create timeline")))?;
+        rendered.push(format!("{}{:>8}", legend, "total"));
+        let timeline = Renderer::new(day_activities.as_slice(), &bounds, &label)
+            .with_renderer(&ActivityRenderer {})
+            .with_length(available_length)
+            .with_boundaries((min_second, max_second))
+            .render()
+            .or_else(|_| Err(anyhow!("failed to create timeline")))?;
+        rendered.push(format!("{}{}", timeline, total_string));
+    }
+    Ok(rendered)
 }