+++ tests/integration_test.rs
@@ -89,6 +89,30 @@ mod tests {
     }
 
     #[test]
+    fn timeline_day_nothing() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("day")
+            .assert()
+            .success();
+    }
+
+    #[test]
+    fn timeline_week_nothing() {
+        let test_dir = tempdir().expect("could not create temp directory");
+        let test_dir_path = test_dir.path().to_str().unwrap();
+        let mut cmd = Command::cargo_bin("rtw").unwrap();
+        cmd.arg("-d")
+            .arg(test_dir_path)
+            .arg("week")
+            .assert()
+            .success();
+    }
+
+    #[test]
     fn continue_none() {
         let test_dir = tempdir().expect("could not create temp directory");
         let test_dir_path = test_dir.path().to_str().unwrap();