+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 * add `day` subcommand (display timeline for the current day)
+* add `week` subcommand (display timeline for the current week)
 
 ## [1.1.0](https://crates.io/crates/rtw/1.1.0) Mar 22, 2020